Rename hardware-specific files for easier tabbing, polish readme
