+++ README.md
@@ -28,7 +28,7 @@ The large text is the current price, the top right numbers are the next three sl
 - You will need *git* and *python 3* available. These are normally installed with Raspberry Pi OS anyway.
 - You will need the Python libraries for the two displays, https://github.com/pimoroni/blinkt and https://github.com/pimoroni/inky
   You can install these libraries directly by entering the following at the Pi command line. Always check that you trust the source of any library (these commands are direct from Pimoroni who make the hardware) and answer **YES** to the questions.
-```
+```sh
 curl https://get.pimoroni.com/blinkt | bash
 curl https://get.pimoroni.com/inky | bash
 ```
@@ -36,7 +36,7 @@ curl https://get.pimoroni.com/inky | bash
 ## How to get this code
 Once you have installed the Blinkt! software as above, the easiest way to download this software is to copy and paste the following command, which will make a copy of all the files in a folder called **agile-pi-ndicator** in your home directory. This won't work unless you've installed the libraries above (or installed `git` yourself).
 
-```
+```sh
 cd ~ && git clone --depth 1 https://github.com/nanoamp/agile-pi-ndicator.git
 ```
 
@@ -63,7 +63,7 @@ L = South West England
 M = Yorkshire
 
 You should initially run manually to check everything works, replacing X with your DNO region letter:
-```
+```sh
 ./store_prices.py --region X
 ```
 
@@ -71,30 +71,30 @@ The code will tell you what it's doing and whether it worked. You can run this a
 
 To update the display, run:
 
-```
+```sh
 ./update.py
 ```
 
 This will also tell you what it's doing. If you want to see all the colours available on a Blinkt! display, you can run 
 
-```
+```sh
 ./update.py --demo
 ```
 
 If it's all a bit much, you can blank the display:
 
-```
+```sh
 ./clear.py
 ```
 
 ## Running automatically
 I really can't be bothered to make a systemd timer/service for this. `cron` is so much easier!
 I've included a script to install the cron jobs listed below. Run it like this, replacing X with your region:
-```
+```sh
 ./install_crontab.sh X
 ```
 You can check it's worked by running `crontab -l`, you should see this:
-```
+```sh
 @reboot /bin/sleep 30; cd /home/pi/agile-pi-ndicator && /usr/bin/python3 store_prices.py --region X > ./pi-ndicator.log 2>&1
 @reboot /bin/sleep 40; cd /home/pi/agile-pi-ndicator && /usr/bin/python3 update_blinkt.py > ./pi-ndicator.log 2>&1
 */30 * * * * /bin/sleep 5; cd /home/pi/agile-pi-ndicator && /usr/bin/python3 update.py > ./pi-ndicator.log 2>&1
@@ -110,7 +110,7 @@ You can check it's worked by running `crontab -l`, you should see this:
 ## Troubleshooting
 
 If something isn't working, run 
-```
+```sh
 less ~/agile-pi-ndicator/pi-ndicator.log
 ```
 This will show you the most recent message from any of the scripts (that were run automatically by `cron`). If this doesn't shed any light, run `./store_prices.py` and `./update.py` and see what they moan about!
@@ -121,17 +121,17 @@ If you want to change price thresholds or fine-tune the colours, they are locate
 
 ### Colours
 Look for this part:
-```
-COLOUR_MAP = { 'level6': { 'r': 155, 'g': 0, 'b': 200 }, # magenta
-              'level5': { 'r': 255, 'g': 0, 'b': 0 }, # red
-              'level4': { 'r': 255, 'g': 30, 'b': 0 }, # orange
+```python
+COLOUR_MAP = { 'level6': { 'r': 155, 'g': 0, 'b': 200, 'name': 'magenta' },
+              'level5': { 'r': 255, 'g': 0, 'b': 0, 'name': 'red' },
+              'level4': { 'r': 255, 'g': 30, 'b': 0, 'name': 'orange' },
 ...
 ```
 **Don't change the names of the levels**. You can change the numbers and the comment after the # sign at the end of the line but nothing else. Any number between 0 and 255 is fine, anything else may produce "interesting" results. 0 means none of that colour, 255 means all of it (modified by the `BRIGHTNESS` setting further up the file). You can test the colours after saving your changes by running `./update.py --demo` and always find the original values on the github site.
 
 ### Price thresholds
 Look for this part:
-```
+```python
     if price > 28:
         pixel_colour = 'level6'
 
@@ -148,6 +148,7 @@ Look for this part:
 
 - better retry if data is late, 3 cron jobs is hacky
 - colour thresholds could depend on daily average rather than fixed values
+- separate log storage and rotation
 
 
 ## Thanks to:
